Add tests and implement various Vm167 class methods
Added several new test methods in `VM167Tests.cs` to test functionalities of the `Vm167` class, including `ResetCounter`, `Connected`, `VersionFirmware`, `ReadBackPWMOut`, and `ReadBackInOutMode`.

Modified the `IVm167` interface in `IVm167.cs` to change the `VersionDLL` method from an asynchronous `Task<int>` to a synchronous `int` return type.

Implemented the `VersionDLL` method in the `Vm167` class to return a hardcoded version value.

Implemented the `ResetCounter` method in the `Vm167` class to reset the counter for a given card address.

Implemented the `Connected` method in the `Vm167` class to check the connection status of devices.

Implemented the `VersionFirmware` method in the `Vm167` class to read the firmware version of a given card address.

Implemented the `ReadBackPWMOut` method in the `Vm167` class to read the PWM output values for a given card address.

Implemented the `ReadBackInOutMode` method in the `Vm167` class to read the input/output mode for a given card address.

Modified the `Write` method in the `Vm167` class to return the number of bytes written and handle null device cases.

Modified the `Read` method in the `Vm167` class to return the number of bytes read and handle null device cases.