Refactor methods, add logging, and improve thread safety
Renamed `OpenDevices` and `CloseDevices` to `Open` and `Close`, and changed their access modifiers to private. Replaced `Debug.WriteLine` with `_logger.LogInformation` for logging. Introduced `SemaphoreSlim` for concurrency control and added `SemaphoreSlimExtensions` for easier usage. Added `ILogger<Vm167>` instance for logging in `Vm167` class. Renamed `NumPWMOut` to `NumPwmOut` for consistency. Updated various methods to use `SemaphoreSlim` and include logging. Changed `ScanPort` to an instance method and updated it to use `_logger` for logging.