Remove SetCard method and related EventToCommandBehavior
The `SetCard` method in the `MainViewModel` class has been removed, along with its `[RelayCommand]` attribute. This method was responsible for logging trace messages when it was called.

In the `MainPage.xaml` file, the `EventToCommandBehavior` that was associated with the `SetCardCommand` for the `RadioButton` elements with content "0" and "1" has been removed. This behavior was used to execute the `SetCardCommand` when the `CheckedChanged` event was triggered for these `RadioButton` elements.